requirements.txt
python-dotenv==1.0.1
aiogram==3.4.1
firebase-admin==6.5.0